Recommit - I thought I did this before, but it was not showing up in the repository.
Makefile.in, doc/Makefile.in: Modified so that making the docarchive
now works.  The make must be done from the top level for it to work, since
some values are passed, but the doc/Makefile.in was missing some declarations
and had some other fixes that needed to be done.  msw 3/22/2000


git-svn-id: https://svn.code.sf.net/p/crossfire/code/trunk/crossfire@203 282e977c-c81d-0410-88c4-b93c2d0d6712