My first commit to sourceforge.. :-))
Fixing a real nasty bug: When a player dies on
battleground (combat arena), he used to loose 1 luck.
Fixed, but: This luck-penalty still applies for
killing another player with melee outside of an arena.
That is not a bug, it is intended.. However, it is not
really a good way to deal with potential "PKers" IMHO.


git-svn-id: https://svn.code.sf.net/p/crossfire/code/trunk/crossfire@821 282e977c-c81d-0410-88c4-b93c2d0d6712