My first commit to sourceforge.. :-))
Fixing a real nasty bug: When a player dies on
battleground (combat arena), he used to loose 1 luck.
Fixed, but: This luck-penalty still applies for
killing another player with melee outside of an arena.
That is not a bug, it is intended.. However, it is not
really a good way to deal with potential "PKers" IMHO.