Apply the big luck patch to all these files.  This removes the special cases for alchemy and prayer as well.
Discussed at length with dhb, av and gros.  Has been tested on both the
mids crossfire server, and on my private server.

This code removes all current use of the luck stat for players, and replaces
it with a general luck modifier for all random rolls.  The intent of this
code is to give a general positive (or negative) effect on a player due
to luck.  This effect should be generally intangible to the player, yet
help him out over the course of a game.

garbled 2001-09-26


git-svn-id: https://svn.code.sf.net/p/crossfire/code/trunk/crossfire@1294 282e977c-c81d-0410-88c4-b93c2d0d6712