Added friendly_fire to attack.c, called from hit_player.
Will put a cap on maxdam that is a percentage (1-100%)
of damage (min 1)based on SET_FRIENDLY_FIRE value (in settings)when
players attack other players in peaceful mode.  Changed who, statistics
commands to show this setting better and added a help file for peaceful.

It would be nice to have a indicator (even a toggle button?)in the
client for this as well.


git-svn-id: https://svn.code.sf.net/p/crossfire/code/trunk/crossfire@2287 282e977c-c81d-0410-88c4-b93c2d0d6712