Slight modification to smoothing code.
Now it's possible to desactivate the smoothing on a map adding

nosmooth 1

in map header. Some mapmaker asked for a 'special arch' which would
disable smoothing on a whole region of map. I think disabling for a whole
map is easier. Maps don't look crappy because of the special arch visibility
(think about no_spell). So no special arch, just about 4 additionnal lines
of code :P. No changes to protocol, so no change to client!


git-svn-id: https://svn.code.sf.net/p/crossfire/code/trunk/crossfire@2525 282e977c-c81d-0410-88c4-b93c2d0d6712