Slight modification to smoothing code.
Now it's possible to desactivate the smoothing on a map adding

nosmooth 1

in map header. Some mapmaker asked for a 'special arch' which would
disable smoothing on a whole region of map. I think disabling for a whole
map is easier. Maps don't look crappy because of the special arch visibility
(think about no_spell). So no special arch, just about 4 additionnal lines
of code :P. No changes to protocol, so no change to client!