Smoothing. Support in gtk-sdl mode and excluded smoothlevel 0
Added support for smoothing in the sdl rendering routines of gtk client. Support
works good with fog of war and lightning.
Also excluded the smoothlevel 0 of rendering routine. This mean a face having a
smoothlevel of zero will not be put underneath the one with smoothlevel 1 or higher.
This also mean a face with smoothlevel 1 will never go on top of another but
may be partially hidden by other faces with smoothlevel 2 or higher.
Best way to check those changes is too look at the earth (having smoothlevel 0) in
the HallOfSelection