Added logging support and bug report help
Remove lots of printf to stderr and replaced them with a LOG call. LOG
also keep in memory the message, so the client can ask them to help bug
reporting. If you use the gtk client and click on help -> Report a Bug
you will see the list of error message in a textarea so you can cut and
paste in bug tracker.
Also modified the LOG prototype so it accepts as parameters a loglevel,
a location (most of time send <partofclient::functionname> in it), a
format string and parameters used in printf functions.
Here is a typical log after connecting to a server:

MESSAGES TRACK:
[  INFO  ] (Client Version) GTK Unix Client 1.6.0
[  INFO  ] (gtk::init_cache_data) Init Cache
[ ERROR  ] (Library::Gtk-CRITICAL) file gtkbox.c: line 332 (gtk_box_pack_start): assertion `child->parent == NULL' failed.
[WARNING ] (common::VersionCmd) Differing C->S version numbers (1022,1023)
[  INFO  ] (common::VersionCmd) Playing on server type  Crossfire Server

[WARNING ] (common::SetupCmd) Server returned FALSE on setup command sexp
[  INFO  ] (commands.c) addme_success received.

[ ERROR  ] (Library::Gtk-CRITICAL) file gtkbox.c: line 332 (gtk_box_pack_start): assertion `child->parent == NULL' failed.

You may notice the gtk / gdk errors are send to the Logging facilities as well.
I plan to add additionnal of [INFO] in code so in case of bug we may have useful
informations like server connected to, server protocol, and config of client.