Check unique map path before saving
Unique maps are assumed to have had their paths correctly set to a
writable directory in LOCALDIR. This assumption may not hold if a map
maker errs (for example, when a player enters a unique map through an
exit that does not have the unique flag set).

Check that the map path is in LOCALDIR before attempting to save it.
Otherwise, the map may be saved to any directory to which the Crossfire
process has write access, which is surprising at best and a security
vulnerability at worst.

git-svn-id: https://svn.code.sf.net/p/crossfire/code/server/trunk@20696 282e977c-c81d-0410-88c4-b93c2d0d6712