Fix unmake_perma_dead() for logged-in accounts
If an account whose character is being resurrected is currently logged in,
the original function would update the character's isDead attribute on
disk without updating the version in memory. Subsequently, when the
account logs out, the version in memory would overwrite the changes.

Fix by obtaining the character list from memory if an account is
currently logged in.

git-svn-id: https://svn.code.sf.net/p/crossfire/code/server/trunk@20763 282e977c-c81d-0410-88c4-b93c2d0d6712