Stop double-subtracting player speed
Player speed is subtracted in do_server() when commands are processed
from the client. It is subtracted again in process_events(), because
players are just objects. This means that even when players are sitting
around, a command might not be processed immediately because the player
used up all their speed in process_events().

Player objects now process_object() at an effective speed of 1.0, as
long as they didn't consume all of their speed handling commands. This
allows the first command to always be processed immediately and reduces
latency.

git-svn-id: https://svn.code.sf.net/p/crossfire/code/server/trunk@20991 282e977c-c81d-0410-88c4-b93c2d0d6712